Rename MoE paged stash package to ops
